requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.33.13
botocore==1.33.13
Django==3.2.25
djangorestframework==3.15.1
imageio==2.31.2
jmespath==1.0.1
networkx==2.6.3
numpy==1.21.6
packaging==24.0
Pillow==9.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.3.0
s3transfer==0.8.2
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
sqlparse==0.4.4
tifffile==2021.11.2
typing_extensions==4.7.1
urllib3==1.26.19
